+++ CHANGELOG.md
@@ -29,6 +29,8 @@
 - jsonrpc: add verified-by information to `Contact`-Object
 - Remove `attach_selfavatar` config #3951
 
+### Changes
+- add debug logging support for webxdcs #3296
 
 ## 1.106.0